refactor(tokens): extract types to files, use some `natds-themes` colors [DSY-1462]
DSY-1462
