+++ packages/styles/src/tokens/colors/natura/IColorsNatura.ts
@@ -1,88 +1,19 @@
+import { IPrimary } from "./IPrimary";
+import { ISecondary } from "./ISecondary";
+import { IGrayscale } from "./IGrayscale";
+import { ILevel } from "./ILevel";
+import { IComplementary } from "./IComplementary";
+import { IGradient } from "./IGradient";
+
 /**
- * @todo Separate classes from IColorsNatura.ts file
- * @todo Set `max-classes-per-file` to [true, 1]
+ * @deprecated `IColorsNatura` is deprecated since v0.20 and will be removed at v1.0.
+ * Please use `themes.natura.light` or `theme.natura.dark` interfaces.
  */
-export interface IPrimary {
-  colorBrdNatPrimaryOrange: string;
-  colorBrdNatPrimaryYellow: string;
-  colorBrdNatPrimaryBrown: string;
-  colorBrdNatPrimaryWhite: string;
-}
-
-export interface ISecondary {
-  colorBrdNatSecondaryPurple: string;
-  colorBrdNatSecondaryBlue: string;
-  colorBrdNatSecondaryGreen: string;
-  colorBrdNatSecondaryYellow: string;
-}
-
-export interface IGrayscale {
-  colorBrdBlack: string;
-  colorBrdNatGray: string;
-  colorBrdNatGray1: string;
-  colorBrdNatGray2: string;
-  colorBrdNatGray3: string;
-}
-
-export interface ILevel {
-  colorLevelSeed: string;
-  colorLevelBronze: string;
-  colorLevelSilver: string;
-  colorLevelDiamond: string;
-  colorLevelGold: string;
-}
-
-export interface IComplementary {
-  colorBrdNatCpBlue1: string;
-  colorBrdNatCpBlue2: string;
-  colorBrdNatCpBlue3: string;
-  colorBrdNatCpBlue4: string;
-  colorBrdNatCpGreen1: string;
-  colorBrdNatCpGreen2: string;
-  colorBrdNatCpGreen3: string;
-  colorBrdNatCpGreen4: string;
-  colorBrdNatCpYellow1: string;
-  colorBrdNatCpYellow2: string;
-  colorBrdNatCpYellow3: string;
-  colorBrdNatCpYellow4: string;
-  colorBrdNatCpOrange1: string;
-  colorBrdNatCpOrange2: string;
-  colorBrdNatCpOrange3: string;
-  colorBrdNatCpOrange4: string;
-  colorBrdNatCpRed1: string;
-  colorBrdNatCpRed2: string;
-  colorBrdNatCpRed3: string;
-  colorBrdNatCpRed4: string;
-  colorBrdNatCpBrown1: string;
-  colorBrdNatCpBrown2: string;
-  colorBrdNatCpBrown3: string;
-}
-
-export interface IGradient {
-  colorBrdNatGrad1: string;
-  colorBrdNatGrad2: string;
-  colorBrdNatGrad3: string;
-  colorBrdNatGrad4: string;
-  colorBrdNatGrad5: string;
-  colorBrdNatGrad6: string;
-  colorBrdNatGrad7: string;
-  colorBrdNatGrad8: string;
-  colorBrdNatGrad9: string;
-  colorBrdNatGrad10: string;
-  colorBrdNatGrad11: string;
-  colorBrdNatGrad12: string;
-  colorBrdNatGrad13: string;
-  colorBrdNatGrad14: string;
-  colorBrdNatGrad15: string;
-  colorBrdNatGrad16: string;
-  colorBrdNatGrad17: string;
-}
-
 export interface IColorsNatura {
-  primary: IPrimary;
-  secondary: ISecondary;
-  grayscale: IGrayscale;
-  level: ILevel;
   complementary: IComplementary;
   gradient: IGradient;
+  grayscale: IGrayscale;
+  level: ILevel;
+  primary: IPrimary;
+  secondary: ISecondary;
 }
