+++ packages/styles/src/tokens/colors/natura/IComplementary.ts
@@ -1,0 +1,29 @@
+/**
+ * @deprecated `IComplementary` is deprecated since v0.20 and will be removed at v1.0.
+ * Please use `themes.natura.light` or `theme.natura.dark` interfaces.
+ */
+export interface IComplementary {
+  colorBrdNatCpBlue1: string;
+  colorBrdNatCpBlue2: string;
+  colorBrdNatCpBlue3: string;
+  colorBrdNatCpBlue4: string;
+  colorBrdNatCpGreen1: string;
+  colorBrdNatCpGreen2: string;
+  colorBrdNatCpGreen3: string;
+  colorBrdNatCpGreen4: string;
+  colorBrdNatCpYellow1: string;
+  colorBrdNatCpYellow2: string;
+  colorBrdNatCpYellow3: string;
+  colorBrdNatCpYellow4: string;
+  colorBrdNatCpOrange1: string;
+  colorBrdNatCpOrange2: string;
+  colorBrdNatCpOrange3: string;
+  colorBrdNatCpOrange4: string;
+  colorBrdNatCpRed1: string;
+  colorBrdNatCpRed2: string;
+  colorBrdNatCpRed3: string;
+  colorBrdNatCpRed4: string;
+  colorBrdNatCpBrown1: string;
+  colorBrdNatCpBrown2: string;
+  colorBrdNatCpBrown3: string;
+}
