+++ packages/styles/src/tokens/colors/natura/IGradient.ts
@@ -1,0 +1,23 @@
+/**
+ * @deprecated `IGradient` is deprecated since v0.20 and will be removed at v1.0.
+ * Please use `themes.natura.light` or `theme.natura.dark` interfaces.
+ */
+export interface IGradient {
+  colorBrdNatGrad1: string;
+  colorBrdNatGrad2: string;
+  colorBrdNatGrad3: string;
+  colorBrdNatGrad4: string;
+  colorBrdNatGrad5: string;
+  colorBrdNatGrad6: string;
+  colorBrdNatGrad7: string;
+  colorBrdNatGrad8: string;
+  colorBrdNatGrad9: string;
+  colorBrdNatGrad10: string;
+  colorBrdNatGrad11: string;
+  colorBrdNatGrad12: string;
+  colorBrdNatGrad13: string;
+  colorBrdNatGrad14: string;
+  colorBrdNatGrad15: string;
+  colorBrdNatGrad16: string;
+  colorBrdNatGrad17: string;
+}
