+++ packages/styles/src/tokens/colors/natura/IGrayscale.ts
@@ -1,0 +1,11 @@
+/**
+ * @deprecated `IGrayscale` is deprecated since v0.20 and will be removed at v1.0.
+ * Please use `themes.natura.light` or `theme.natura.dark` interfaces.
+ */
+export interface IGrayscale {
+  colorBrdBlack: string;
+  colorBrdNatGray: string;
+  colorBrdNatGray1: string;
+  colorBrdNatGray2: string;
+  colorBrdNatGray3: string;
+}
