+++ packages/styles/src/tokens/colors/natura/ILevel.ts
@@ -1,0 +1,11 @@
+/**
+ * @deprecated `ILevel` is deprecated since v0.20 and will be removed at v1.0.
+ * Please use `themes.natura.light` or `theme.natura.dark` interfaces.
+ */
+export interface ILevel {
+  colorLevelSeed: string;
+  colorLevelBronze: string;
+  colorLevelSilver: string;
+  colorLevelDiamond: string;
+  colorLevelGold: string;
+}
