+++ packages/styles/src/tokens/colors/natura/IPrimary.ts
@@ -1,0 +1,10 @@
+/**
+ * @deprecated `IPrimary` is deprecated since v0.20 and will be removed at v1.0.
+ * Please use `themes.natura.light` or `theme.natura.dark` interfaces.
+ */
+export interface IPrimary {
+  colorBrdNatPrimaryOrange: string;
+  colorBrdNatPrimaryYellow: string;
+  colorBrdNatPrimaryBrown: string;
+  colorBrdNatPrimaryWhite: string;
+}
