+++ packages/styles/src/tokens/colors/natura/ISecondary.ts
@@ -1,0 +1,10 @@
+/**
+ * @deprecated `ISecondary` is deprecated since v0.20 and will be removed at v1.0.
+ * Please use `themes.natura.light` or `theme.natura.dark` interfaces.
+ */
+export interface ISecondary {
+  colorBrdNatSecondaryPurple: string;
+  colorBrdNatSecondaryBlue: string;
+  colorBrdNatSecondaryGreen: string;
+  colorBrdNatSecondaryYellow: string;
+}
